Try to spot stuck inputs
If an input causes an error, all future input is ignored until
the user presses <esc> or <ctrl-c>. This commit spots this situation and
clears the current input queue instead.

This is overzelous! Something like 'fy' is a movement that only works
for some cursors and errors for others. This change means all cursors
ignore the movement if the first cursor doesn't have a 'y' on this line.

Still an improvement over the old state of all cursors being stuck and
all future input being ignored, though.

There is a tradeoff here between soundness and performance. We could
also test more, or even all, cursors. This becomes quickly *painfully*
slow, as in several seconds when you have dozens of cursors and complex
commands since these commands might be executed many times per cursor to
test all the different heuristics. Extracting the heuristics into a rule
engine might help to do smarter stochastic sampling instead.

But since I haven't touched this code in a while and there are SO MANY EDGECASES I
will leave it at good enough for now.