Work around neovim bug
Complex motions like `/` should be allowed to be interactive. The
battle plan is:

- setup a fake operator
- `call feedkeys('g@v/', 'ix!')` to give control to the user via the `!`
  flag
- repeat the operator with `.` to repeat the motion

The feedkeys setup used to be in `multi#util#apply_op`.

But! Neovim cleans up the repeat register when returning from functions!
I consider this a bug - it used to be broken, then worked, and now seems to be broken again.
Work around this by inlining `multi#util#apply_op` and ignore how silly this
all is.

There might be other places where the same bug still waits.